Use original posterior
This posterior is the one provided by Dalton. Rhats are not all <1.01,
the model isn't really converged. Would be nice to combine these samples
with the ones I've got in data/post2016.rds, but it needs another 500
samples to be able to line up (2k samples each chain) as four new
chains, which should improve things. Merging as a single long chain
gives good Rhats, but this isn't quite as rigorous as considering the 8
chains separately.